+++ app/src/main/java/com/pinnacleimagingsystems/ambientviewer/Algorithm.java
@@ -4,11 +4,11 @@ public interface Algorithm {
     interface Meta {
         int parameterMin();
         int parameterMax();
-        int defaultParameter(int lux);
+        float defaultParameter(int lux);
     }
 
     Meta getMeta();
 
-    void init(int parameter);
+    void init(float parameter);
     void apply(int[] rgbaPixels, int width, int height);
 }
